[Network Identification Test]
Started on making a handshake routine to let the switch broadcast itself on the network.

PC will hear the broadcast and respond with the key that was in the cast.

Switch will confirm the key matches, and if so store the IP of the PC for later connections.

Need to properly plan out the send and receive patterns for the PC and switch. Need to write the PC application to accept parameters from the switch and use them for the ffmpeg stream execution. Also need to make the applicaiton execute the stream.