[Video Stream - Start/stop stream without restart]
Cleaned up main and moved all test/experimental code into a text file for later reference.

Placed input on a separate thread, with some duct taped mutexes to get it working for now. Need to really sit down and design the control flow for the input, renderer, and streamer.

Added the ability to let the switch kill the video stream now (holding 'plus')

The switch will close the stream, and can then the user can hold 'R' to let the switch wait for another stream connection. And it connects again without crashing! :D

Attempted rendering on a thread other than Main, SDL didn't like that. So now, anything that needs to not be blocked by the renderer needs to run on an independent thread.

Watching the CPU load via Status Monitor showed that the child thread is placed on the same core as Main. Suggests that if the video stream class were re-written to be executed iteratively, it could function comfortably with input polling on the same thread. Something to consider.

Weird latency issues tonight. Could be from the input thread sleeping and waking up on the same core as the stream. Need to confirm.