[Latency Experiments - FFMPEG args]
Removed locks on sections that weren't critical to minimise unnecessary stalling of the stream thread. The stream currently runs well on default clock speeds.

Made a text file to keep track of the many ffmpeg argument configurations I've tested.

Experimented in reducing stream latency with FFMPEG args.

The '-vsync' arg seems to be the most important one to define for stream performance.