[Network Payload Commands PoC]
Made some test code to see how sending payload data from the switch to the server could work.

Will next need to be able to make a dummy client to test the server functionality for configuring ffmpeg, starting the stream, and stopping it. Then implementing all the functionality into the server.

Once functionality is tested and a working prototype is made, the switch version of the client will start. After that is working, then the server application will be decoupled from this project and moved into a separate one to reduce the horrible clutter and intellisense issues occurring in VS Code.