[Start Stream Signal PoC]
Added the ability to send command codes over network from switch to PC.

Added the ability to start the stream on the PC from the switch via button press.

Note: the command connection code crashes the application and sometimes atmosphere if started in another thread. The crash occurs right after SendCode is called and succeeds. For now, the connection is performed by the main thread to ask the PC to start the stream, and then it attempts to connect to the stream after (this however adds a little bit of latency).