[Switch Remote Play Host - Windows Project]
Finally got around to adding the C++ project I've been making to act as the stream host. Became paranoid of data loss and I really don't want to re-write all that.

Was considering hosting the windows project on a separate repo but I wanted to keep the projects and their dependencies localised so it is easy to setup and modify on a different machine. This includes the build ViGEmClient libraries and the version of FFMPEG I'm using on Windows. (Building FFMPEG is quite tedious on Windows...)

Setup the gitignore files for the ExtraDependencies folder and the windows project folder. Need to keep an eye out for this.