[Automatic Network Discovery - Switch <--> PC]
The Switch and Windows application now handshake and discover each other. This took quite a lot of testing. Mainly to allow the program to terminate without being hung up due to a socket waiting.

The host PC IP address is displayed on the screen when it is found. The next things to look into are using fewer ports, and discovering the local subnet. That is one of the few things left hardcoded in for network connectivity. (Currently assumes subnet is 192.168.0.255)

Added the ability to shutdown sockets used in Broadcast and Connection.
This allowed blocking sockets to be stopped, and let the program terminate.

Fixed a bug with Broadcast's close function not closing the receiver socket.

Lastly... It works! Yay! :D