[Network Discovery - Stage 1]
Network Discovery now is performed on button press.

Currently a reference to a pointer managed by the input thread is used. This will be replaced when a copy constructor and copy assignment operator is made.

Stage 2 will be designing a paged UI system and having the UI handle the input and finally refactoring away the InputThread.h file and moving the inputs away from being on a separate thread.