[Host Listening + Default FFMPEG]
Modified the host application to have a background thread that listens for broadcasts from a switch. This allowed the host to not have to be restarted everytime the switch application is restarted.

Modified the host application's Connection class to be in parity with the switch version.

Modified the default FFMPEG settings on the switch UI to be settings with minimal latency.