[FFMPEG UI Rollback + Main cleanup]
Rolled back the implementation of the FFMPEG UI. New implementation had issues with reference capturing in the lambda callback functions.

Cleaned up main by removing the use of the separate input thread when the stream was inactive. Now the input is simply polled within main and has significantly reduced the boiler plate code needed due to the input thread from previous.

Moved the rendering of the main screen when the stream is inactive from main to a separate code file.