[Monitor Resolution Adjustments]
Added the ability to change the monitor resolution to match the chosen video resolution from the FFMPEG settings.

Added a check for when the connection to the gamepad thread on the switch closes so it can force the stream to stop. However, if the switch goes into airplane mode (killing network connectivity) the gamepad thread is not notified until the switch reconnects to the network again. Meaning the stream cannot be cleaned up if the network connection drops unexpectedly.

Added more 16:9 resolution options and bitrate options in the switch app. Also added better explanations for the vsync modes.