[Audio Stream + New Crash]
-- Bug --
Program crashes now when opened twice within the same HBLoader instance. (Open app, exit the app. Open the app again and HBLoader will crash)

-- Audio --
Added the ability for audio to be streamed via FFMPEG to the switch.

The audio playback code is based from In-Home-Switching code, with some replacements for constants and minor cleanups to be able to repeatedly call the playback code. A shell class is defined at the moment to replace the C code after refactoring.

Need to reimplement the audio playback code. Also need to figure out what the intermittent popping sound is caused from.