[Audio Re-implementation - Configurable Audio]
Figured out how to allow audio to be configurable now. However, is not implemented yet. The object to allow for it to be so is already done. (See PcmStream.h/cpp)

Thanks to the playtone switch-example I could play around with it and determine what was going wrong in the previous implementation. The short of it is I wasn't waiting on the framerate's worth of packets before attempting to play the buffer.

Now, the framerate can be adjusted, the sampling rate, and even the bitrate too. The bitrate however needs a special copy operation if it is to be configurable (unless the hardware bitrate can be changed or an audio device with different bitrates is available. Haven't checked yet).

Current implementation avoids many copy operations and has a reduced memory footprint compared to the AudioStream implementation.

AudioStream was based off of the In-Home-Switching audio.c file and will be deleted or re-purposed after further testing.

Currently a softlock can occur if the audio stream is closed from the PC.