[System Font Crash Fix + Sleep Audio Fix]
Added some TrueType fonts to the project to be used for the UI.

The application crash was occurring when trying to load the shared fonts from the system. No longer there with fonts being loaded from romfs.

Fixed a sleep issue where the audio stream would fail to restart correctly if the stream died due to the switch sleeping.

Added the ability to redirect stdout to a file. HOWEVER, it causes any switch application to crash when opened after this application unless the hbmenu is closed. My guess is it is due to the stdout not being restored.