[Audio Refactor + Code Cleanup]
Refactored out the audio buffering and playback into the AudioPlayback class.

Added an AudioConfig struct to be sendable to the host application and used for setting up audio playback.

Deleted references to old audio implementation.

Deleted unnecessary code from VideoStream and prepared it to have the port configurable.

-- BUG --
A new crash is occurring trying to load the system fonts from romFS when the application is loaded twice after using a stream in the first instance of the application opening.