[Menu Screens - Config and IP]
Added some logic to display many menu screens. For now there is a main, config, and IP screen.

The main screen just displays the current configuration for now.

The config screen is the config UI to allow users to change the ffmpeg settings.

The IP screen is to allow users to manually set the IP for the host. Only renders for now and the displayed manual IP is not used.