[Host Application - Port Assignment]
Matched the port declerations in the switch application with the PC application. Later need to make the ports loadable from a file. Also need to make the IP found in the broadcaster thread the IP used for the video and audio stream started by the PC application.

Cleaned up unused code in the input thread file for the switch application.