[Manual IP Feature + UI Tweaks]
Added saving manual IP to the config file.

Allowed users to toggle and use manual IP defined in the manual IP screen.

Modified setting buttons to be A and B instead of the D-Pad.

-- UI tweaks --
Removed unused fonts.

Changed current font to a monospaced font.
Changed font size to compensate.

Repurposed the MAIN screen to be a help screen.

Moved around and re-positioned other text elements.

-- windows project - host application --
Modified host application to be able to connect when the switch manually established a connection instead of being discovered and doing a handshake beforehand.