[Mouse Input Prototype + Cleanup]
Added mouse input to the host application to allow users to move the mouse, left click, and right click.

Mouse and controller input mode is toggled by holding a button combo for greater than 3 seconds.

Auto minimises the console window if it is shown when the stream starts.

Added a version string to help identify which host application is running.

-- Cleanup --
Cleaned up the main.cpp header declarations.

Removed the option to set switch scale res to 1920 x 1080. YUV420 stride needs to be calculated for it correctly.

Marked the area where a very bad bug/crash occurs when the application has been closed after a stream.

Modified the socket shutdown functions for the Switch's Broadcast/Connection.cpp files to close the socket after shutting down.

Deleted unused files and old code files left from the prototyping phase.

Added an extra tip to the help screen explaining mouse and controller toggle.

-- Bug --
Updated readme to explain the current issue the application creates not just for itself, but for homebrew loader as well.