[Menu Refactor + ffmpeg pathing]
--- Windows Project ---
Modified windows project to use the executable path passed in the main args to determine parent directory.

--- Switch Project ---
Refactored the configuration screen and network menu screens to use the Menu interface.

Menu was refactored to be a simple interface for implementing concrete menu classes.

More refactoring is necessary in order to include 2 upcoming additional screens to allow user configuration.