[Menu and Enum Refactor]
--- Menu Screen Refactor ---
Refactored the MainScreen.h/cpp files to become a menu selection menu.

Refactored all screens to implement the Menu interface. Now adding extra menus can be done by implementing the interface and adding it as an entry in the MenuSelection class.

--- Enum Refactor ---
Refactored all enums to their own h/cpp files. Including parser functions to help for display, saving, and loading.

Deleted all re-definitions of the parser functions from Configuration and from ConfigurationScreen.

Updated the NetworkData header to correct the VsyncMode variable to be an enum now.

--- Windows Project ---
Updated to mirror the enum updates.