[Decoder Menu]
Added a decoder menu that allows users to configure the decoder settings on the Switch.

Not all flags are made available and not all available flags necessarily will have any effect. Only if the portlibs ffmpeg library has the related flag features compiled into it. However they are there for both placebo and for if the ffmpeg portlib gets updated to include these flag features.