[Controller Configuration Menu + DS4 Controller Bug Fix]
Updated the controller configuration menu to allow users to configure face button mapping, mouse sensitivity, mouse on connect option, and controller choice.

--- Windows Project ---
Fixed the DS4 Bug introduced with the previous pull request. The DPad state wasn't being updated in the correct order.

Refactored the Virtual Controller interface as well as the X360 and DS4 implementations.

Removed button mapping hotkey and replaced it with a map function as part of the initialisation process.

Replaced raw pointer of VirtualController with a unique pointer.