[PC Diagnostic Info - Output Initial State]
Added extra diagnostic information that gets printed at the startup of the application.

Basic monitor info is displayed.

Network state with important address info is displayed. If any info is missing or no active networks are found, then a warning is displayed.

A sanity check of the virtual controller driver is done at the startup now to confirm the driver is working. If it fails then instructions to reinstall are displayed.

On stream connection, if the gamepad thread starts up and fails to create a controller it now outputs an error informing the user why the stream was closed.