[Controller Bug Fixes + Ffmpeg Encoder Logging Options]
Added a configuration file that allows the control of displaying the ffmpeg program logging output data. This will help with collection diagnostic info.

Fixed the DS4 and 360 controllers to use the constants for data type limits rather than hardcoding values.

Fixed the Switch's minus button being mapped to the 360 'guide' button. Now it correctly maps to the 'back' button on the 360 controller.

Fixed a socket read bug in the controller stream. If a network card's maximum MTU is smaller than the size of the game pad payload, rest of the payload would be initialised with garbage data causing bad inputs. Changed now to loop on the socket and ensure that the entire payload is read before proceeding to simulate controller inputs.