[Encoder codecs - h264(amf) definitions]
Added the codec definition, configuration, and utility files that were being worked on in a separate test project.

Next is to cleanup header guards/references and then start refactoring the UI to use the new classes.

Then the old configuration class needs to be repurposes into a 'main configuration' class to handle data not related to encoders.