[Version 0.9.0 - Documentation Update]
Deleted the old instructions and documentation.

Modified the readme to instruct users to go to the wiki or the gbatemp thread for help.

Added some basic markdown files for the new instructions for version 0.9.0.

The main importance are the images, so the wiki can reference them in the documentation.

Rebuilt the ViGEm Client libraries since they were compiled with an older visual studio compiler.