[Touch Support - True Single Finger Touch] #17
Implemented basic virtual touch interface that works with 1 finger for now.

The touch input is reported as a true touch input to Windows. This allows the Switch to interact with Windows as if it were a Windows tablet.

5 finger support along with multi-finger ID tracking needs to be implemented.

Simulated virtual mouse via touch input also needs to be implemented to allow for users to choose to use the touch screen as just a mouse instead of a touch screen in Windows.