[Touch Support - Multi Touch Prototype] #17
Modified virtual touch class to work with multiple fingers.

One finger interactions work like previous implementation.

Two finger interactions do not work as if the 2nd finger was late (despite being injected at the same time). Leaving only 1 finger registering as being pressed down while the 2nd is registered as moving.

3 or more causes 2 'ghost' fingers to be stuck in a held down state. This can be seen using a touch screen tester.

When 3 or more fingers are placed, they will register movement and press state correctly.

Lost finger tracking and 'swapping' fingers causes all fingers to be released.

Touch polling in the Switch app was added to give a chance to combine touch events that occur in close proximity and to allow for a small buffer window for 2 or greater finger touch to be registered simulaneously on press.