[Monitor Select - Basic Prototype] #21
Modified the display device service class to enumerate monitor info and then map these when enumerating the active display devices. This is used to capture location and resolution data of the display device.

Implemented a prototype for monitor selection. Ffmpeg now can get x and y offset data in order to target a specific monitor.

Modified the Switch application to include a 'monitor number' UI element in the general codec settings menu. This might be replaced since it uses the bugged NumberRange iterator.

Modified the network payload to include a monitor number in the VideoData struct. Switch and Host applications were updated.

Resolution changes are targeted on the selected monitor number.