[Touch Support - Simulated Touch Mouse Prototype] #17
Created a prototype for simulating mouse input via touch states. This still needs to be integrated and tested with the Switch app.

Instead of injecting the touch data as a touch device to windows, the touch states are interpreted and used to simulate mouse events.

This only supports the primary monitor, so monitor selection is not available for this touch input mode.

Feature might be removed as touch injection is prefered. Need to investigate Cemu interaction to see touch injection vs mouse simulation.