[Touch Support - Simulated Touch Mouse Deadzone] #17
Added some tests for implementing a deadzone radius to the simulated touch mouse. Works similarly to the VirtualTouch interface.

Added a utility header to handle deadzone checking. VirtualTouch interface needs to be refactored to use the new utility class.

VirtualMouse interface could potentially use the deadzone utility to limit flooding windows with send input messages for move events.

Side note: tested virtual touch interface with CEMU and it works just fine. So this simulated touch mouse will need to be tied to a separate issue as well as the absolute mouse movement implementation.