[PC Host - Virtual Touch Screen Scaling] fixes #37
Virtual touch input co-ordinates now are scaled based on the active session desktop resolution.

Prior to this, if the desktop resolution was greater than the Switch's screen resolution then the user could only touch the desktop in a 1280 x 720 region. Also if the desktop resolution were lower than the Switch's resolution, the Switch would touch outside the monitor's region.