[Home Button - Button Trigger Prototype] #24
Added the ability to trigger the 'home' button when a key combination is held.

Holding the Left Analog Stick Button for more than 1 second triggers the DS4 controller's 'PS' button or the Xbox360 controller's 'Guide' button respectively.

This will later be made configurable in the Switch application under the controller menu as a button check combo feature.

The Switch app was modified to no longer signal the '_Palma' button input when the touch screen is pressed.