[Virtual Keyboard - Prototype Bindings] #38
Created a basic prototype class for virtual keyboard input. This allows the Switch's buttons to perform keyboard inputs while in 'mouse mode'.

Basic key inputs work, but there is an issue with stuck keys. The hoy key for task manager for example has 'ctrl' and 'shift' remain stuck, despite being sent correctly and released. The keys will not release until the physical keys ctrl, shift, and escape keys are pressed and released on the PC's keyboard.

Simple keys like the arrow keys, enter, left windows, and the backspace keys work fine and are mapped. The modifier keys seem to cause issues and need more investigation considering the 'SendInput' Windows function is not throwing an error either.