[Virtual Keyboard - Long Press Feature] #38
Modified the polling function for virtual keyboard input to track the hold time for a key. This way, it now will function similar to a regular keyboard with the first press of the key registering as a single press, then after a delay becomes repeated presses.

The keys getting stuck that was mentioned in the previous commit was a Windows security feature. Programs sending input to a program with higher elevation will be blocked and fail. So the hotkey to open Task Manager would not be able to release because once it is opened, it is focused immediately. This caused the release events to become blocked by Task Manager.

The remedy to the key input privilege failure, is to run the program as Administrator. Wiki will need to be modified to explain this.

Side note (#27): mouse input has now been modified to have a poll time too, to make sure the mouse movement and scrolling is consistent with time rather than loop speed. This also makes it less likely for flooding mouse input.