[Refactor - Numbers via Software Keyboard] #40
Implemented a templated software keyboard method to allow opening the keyboard to accept and parse user input.

User input can be validated as well via a predicate.

Some number based settings have been modified to use the new software keyboard method now. This can be seen for Thread Count and Mouse Sensitivity.

Some menus will need to be modified to be able to use the software keyboard and the network configuration menu needs to be overhauled.

Max thread count for the decoder menu was decreased to a sensible(but still high) value to stop users from completely crippling performance unknowingly.