[PC Host - Mapped Configuration Data]
Modified the configuration class in the PC Host to match the one used in the Switch.

The ffmpeg output options were implemented via the mapped configuration class. Now future configuration options can be implemented following the same pattern that is used in the Switch.

Next up is to include the network options, and some options for on connect behaviour.