[Virtual Keyboard - Keybinder Prototype] #38
Added the ability to allow the user to create key bindings using a button watcher. This allows users to use button combos for key bindings.

A configuration object needs to be made and maybe a second look at how to deal with the render function for the button watcher function.