[Touch Support - Simulated Trackpad Prototype] #17
Implemented the 'simulated mouse mode' from the touch configuration menu.

Discovered that the touch screen could be used as a simulated touch/track pad like a laptop.

Single-tap, double-tap, double-tap and drag, and two finger right-click have been implemented.

An 'absolute' simulated pad could be done too, where the trackpad extends map to the desktop extents.

A virtual interface for 'virtual touch devices' needs to be made and the VirtualTouch class might need to be refactored to use the new interface.