[Touch Menu - Extra Simulate Mouse Options] #17
Added extra configuration options to allow users to choose the desired mouse behaviour for the simulated touch mouse.

Users can choose for the touch screen to behave as a trackpad or as an absolute screen position.

The double-tap time has been made configurable as well.

New configuration options are sent over to the PC to be used when creating the mouse interface.

Next need to clean up the mess of the different mouse implementations and turn them into a virtual class.