[Refactor - Common Code Folder] #43
Refactored Switch application to reference the 'common' folder for code that could be shared between different projects.

Moved out all files that aren't exclusively dependent on the Switch application into a 'common' folder. This is to remove the large amounts of code duplication between the Switch and Windows projects.

xxxxxCodec related objects that handled parameter modifications were modified to take a delegate function. This was so the dependency on the software keyboard was removed.