[Logging - Logging Importance Filter] #43
Implemented a basic logging class to help filter stream output to the console. This will later be configurable to allow for choosing importance threshold and whether or not to output to the console or to a file.

Replaced most areas that printed via cout to now use the log class to output text.

Implemented a simple string/wstring interface to allow the code sites using the log object to not have to worry about string typing.

Added utf8/utf16 string shortener/widener functions.

Added string util functions for concatenation of mismatching string length types.

Timestamp feature needs to be implemented.

Synchronisation primitives might need to be introduced for the logger.