[Refactor - General Config Options] #43
Repurposed the original 'config.ini' file to be used for general program options.

Users can configure the window so it can minimise on connection and restore on disconnection.

Users can configure the PC's audio so it can mute on connection and restore original audio state on disconnection.

Bumped PC program version up to 1.0.0.