[Logging - File Logging] #43
Modified the log class to allow for logging to a text file as well as the console.

Log class needs to be modified now to change the pipe operator to return the log class instead of the underlying stream in order to allow for logging to the console and a file.

Added a general configuration option to turn file logging on or off.