[Audio - Buffer Size Tests] #62
Increased sample size for audio spec setup to better match the Switch's audoi buffer 8192 byte alignment.

Fixed issue with 'choppy' audio by increasing sample size (to make SDL act as if it is double-buffering)

Fixed an issue where the app would fail to close when SDL_Quit was called due to the audioStream not shutting down.